+++ test/busbranch/test_bus_branch.py
@@ -3,178 +3,272 @@
 #  This Source Code Form is subject to the terms of the Mozilla Public
 #  License, v. 2.0. If a copy of the MPL was not distributed with this
 #  file, You can obtain one at https://mozilla.org/MPL/2.0/.
-from typing import Set, Union, FrozenSet
+from typing import Set, Callable, Dict
 
 import pytest
+from hypothesis import given
+from hypothesis.strategies import booleans, sampled_from
 
+from test.busbranch.data.end_of_branch_multiple_ec_pec import end_of_branch_multiple_ec_pec
+from test.busbranch.data.multi_branch_common_lines_network import multi_branch_common_lines_network
+from test.busbranch.data.negligible_impedance_equipment_basic_network import negligible_impedance_equipment_basic_network
+from test.busbranch.data.simple_node_breaker_network import simple_node_breaker_network
+from test.busbranch.data.single_branch_common_lines_network import single_branch_common_lines_network
+from test.busbranch.data.three_common_lines_network import three_common_lines_network
 from test.busbranch.test_bus_branch_creator import TestBusBranchCreator, create_terminal_based_id
-from zepben.evolve import ConnectivityNode, Junction, Disconnector, BusbarSection, Switch, ConductingEquipment, AcLineSegment, BusBranchNetworkCreationMappings
+from zepben.evolve import ConnectivityNode, Junction, Disconnector, BusbarSection, Switch, ConductingEquipment, AcLineSegment, Terminal, NetworkService, \
+    BusBranchNetworkCreationMappings
+# noinspection PyProtectedMember
 from zepben.evolve.model.busbranch.bus_branch import _group_negligible_impedance_terminals, _group_common_ac_line_segment_terminals
 
 
 @pytest.mark.asyncio
-async def test_create_bus_branch_model_callbacks(simple_node_breaker_network):
-    nb_network = simple_node_breaker_network
+async def test_create_bus_branch_model_callbacks():
+    nb_network = simple_node_breaker_network()
     plsi, wire_info, pt_info, es, pt, line, ec, pec = _get_validated_conducting_equipment(nb_network)
+    exp_bb0, exp_bb1, exp_bb2, exp_branch, exp_tx, exp_es, exp_ec, exp_pec = _get_expected(nb_network, line, pt, es, ec, pec)
 
     creator = TestBusBranchCreator()
     result = await creator.create(nb_network)
 
-    was_successful, validator, mappings, bb_network = _get_validated_results(result)
-
-    expected_bus_0, \
-    expected_bus_1, \
-    expected_bus_2, \
-    expected_branch, \
-    expected_transformer, \
-    expected_energy_source, \
-    expected_energy_consumer, \
-    expected_power_electronics_connection = _get_expected(nb_network, line, pt, es, ec, pec)
+    validator, mappings, bb_network = _get_validated_results(result)
 
     _validate_validator_calls(validator)
-    _validate_network(bb_network, expected_bus_0, expected_bus_1, expected_bus_2, expected_branch, expected_transformer, expected_energy_source,
-                      expected_energy_consumer, expected_power_electronics_connection)
-    _validate_mappings(mappings, expected_bus_0, expected_bus_1, expected_bus_2, expected_branch, expected_transformer, expected_energy_source,
-                       expected_energy_consumer, expected_power_electronics_connection, line, pt, es, ec, pec)
+    _validate_mappings(mappings, exp_bb0, exp_bb1, exp_bb2, exp_branch, exp_tx, exp_es, exp_ec, exp_pec, line, pt, es, ec, pec)
+    _validate_network(bb_network, exp_bb0, exp_bb1, exp_bb2, exp_branch, exp_tx, exp_es, exp_ec, exp_pec)
 
 
-def _get_expected(
-    nb_network,
-    line,
-    pt,
-    es,
-    ec,
-    pec
-):
+@pytest.mark.asyncio
+async def test_topological_nodes_inside_grouped_lines_do_not_get_created():
+    nb_network = three_common_lines_network()
+    creator = TestBusBranchCreator()
+    result = await creator.create(nb_network)
+
+    validator, mappings, bb_network = _get_validated_results(result)
+
+    assert len(bb_network.bus) == 2
+    assert len(bb_network.branch) == 1
+
+
+@pytest.mark.asyncio
+@given(sw_is_open=booleans())
+async def test_group_common_ac_line_segment_terminals_single_branch(sw_is_open: bool):
+    nb_network = single_branch_common_lines_network(sw_is_open)
+
+    acls1 = nb_network.get("acls1")
+    acls2 = nb_network.get("acls2")
+    acls3 = nb_network.get("acls3")
+    acls4 = nb_network.get("acls4")
+    acls5 = nb_network.get("acls5")
+
+    await _validate_line_grouping({acls1, acls2, acls3}, {*acls1.terminals, *acls2.terminals, get_term(acls3, 1)}, {get_term(acls3, 2)})
+    await _validate_line_grouping({acls4}, set(), {*acls4.terminals})
+    await _validate_line_grouping({acls5}, set(), {*acls5.terminals})
+
+
+@pytest.mark.asyncio
+async def test_group_common_ac_line_segment_terminals_multi_branch():
+    nb_network = multi_branch_common_lines_network()
+
+    a0 = nb_network.get("a0")
+    a1 = nb_network.get("a1")
+    a2 = nb_network.get("a2")
+    a3 = nb_network.get("a3")
+    a4 = nb_network.get("a4")
+    a5 = nb_network.get("a5")
+    a6 = nb_network.get("a6")
+    a7 = nb_network.get("a7")
+    a8 = nb_network.get("a8")
+
+    await _validate_line_grouping({a0, a1, a2}, {get_term(a0, 1), *a1.terminals, get_term(a2, 1)}, {get_term(a2, 2)})
+    await _validate_line_grouping({a0, a1, a2}, {get_term(a0, 1), *a1.terminals, get_term(a2, 1)}, {get_term(a2, 2)})
+    await _validate_line_grouping({a3}, set(), {*a3.terminals})
+    await _validate_line_grouping({a4, a5}, {get_term(a4, 2), get_term(a5, 1)}, {get_term(a4, 1), get_term(a5, 2)})
+    await _validate_line_grouping({a6, a7}, {get_term(a6, 2), *a7.terminals}, {get_term(a6, 1)})
+    await _validate_line_grouping({a8}, set(), {*a8.terminals})
+
+
+@pytest.mark.asyncio
+async def test_group_common_ac_line_segment_terminals_end_of_branch_multiple_ec_pec():
+    nb_network = end_of_branch_multiple_ec_pec()
+
+    a1 = nb_network.get("a1")
+    a2 = nb_network.get("a2")
+
+    await _validate_line_grouping({a1, a2}, {get_term(a1, 2), get_term(a2, 1)}, {get_term(a1, 1), get_term(a2, 2)})
+
+
+@pytest.mark.asyncio
+async def test_group_negligible_impedance_terminals_single_branch_closed_switch():
+    nb_network = single_branch_common_lines_network(False)
+
+    acls1 = nb_network.get("acls1")
+    acls2 = nb_network.get("acls2")
+    acls3 = nb_network.get("acls3")
+    acls4 = nb_network.get("acls4")
+    acls5 = nb_network.get("acls5")
+    sw = nb_network.get("sw")
+
+    def has_neg_imp(ce) -> bool:
+        if isinstance(ce, Switch):
+            return not ce.is_open()
+        else:
+            return False
+
+    await _validate_term_grouping(has_neg_imp, nb_network, "acls1_acls2", set(), set(), {get_term(acls2, 1), *acls1.terminals})
+    await _validate_term_grouping(has_neg_imp, nb_network, "acls2_acls3", set(), set(), {get_term(acls2, 2), get_term(acls3, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "acls3_sw", {sw}, {*sw.terminals}, {get_term(acls3, 2), get_term(acls4, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "acls4_sw", {sw}, {*sw.terminals}, {get_term(acls3, 2), get_term(acls4, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "acls4_acls5", set(), set(), {get_term(acls4, 2), *acls5.terminals})
+
+
+@pytest.mark.asyncio
+async def test_group_negligible_impedance_terminals_single_branch_open_switch():
+    nb_network = single_branch_common_lines_network(True)
+
+    acls1 = nb_network.get("acls1")
+    acls2 = nb_network.get("acls2")
+    acls3 = nb_network.get("acls3")
+    acls4 = nb_network.get("acls4")
+    acls5 = nb_network.get("acls5")
+    sw = nb_network.get("sw")
+
+    def has_neg_imp(ce) -> bool:
+        if isinstance(ce, Switch):
+            return not ce.is_open()
+        else:
+            return False
+
+    await _validate_term_grouping(has_neg_imp, nb_network, "acls1_acls2", set(), set(), {get_term(acls2, 1), *acls1.terminals})
+    await _validate_term_grouping(has_neg_imp, nb_network, "acls2_acls3", set(), set(), {get_term(acls2, 2), get_term(acls3, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "acls3_sw", set(), set(), {get_term(acls3, 2), get_term(sw, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "acls4_sw", set(), set(), {get_term(sw, 2), get_term(acls4, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "acls4_acls5", set(), set(), {get_term(acls4, 2), *acls5.terminals})
+
+
+@pytest.mark.asyncio
+async def test_group_negligible_impedance_terminals_multi_branch():
+    nb_network = multi_branch_common_lines_network()
+
+    a0 = nb_network.get("a0")
+    a1 = nb_network.get("a1")
+    a2 = nb_network.get("a2")
+    a3 = nb_network.get("a3")
+    a4 = nb_network.get("a4")
+    a5 = nb_network.get("a5")
+    a6 = nb_network.get("a6")
+    a7 = nb_network.get("a7")
+    a8 = nb_network.get("a8")
+
+    def has_neg_imp(ce) -> bool:
+        if isinstance(ce, Switch):
+            return not ce.is_open()
+        else:
+            return False
+
+    await _validate_term_grouping(has_neg_imp, nb_network, "a0_a1", set(), set(), {get_term(a1, 1), *a0.terminals})
+    await _validate_term_grouping(has_neg_imp, nb_network, "a1_a2", set(), set(), {get_term(a1, 2), get_term(a2, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "a2_a3_a6", set(), set(), {get_term(a2, 2), get_term(a3, 1), get_term(a6, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "a3_a4_a8", set(), set(), {get_term(a3, 2), get_term(a4, 1), *a8.terminals})
+    await _validate_term_grouping(has_neg_imp, nb_network, "a4_a5", set(), set(), {get_term(a4, 2), get_term(a5, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "a5", set(), set(), {get_term(a4, 2), get_term(a5, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "a6_a7", set(), set(), {get_term(a6, 2), *a7.terminals})
+
+
+@pytest.mark.asyncio
+@given(nie_constructor=sampled_from([Junction, Disconnector, BusbarSection]))
+async def test_group_negligible_impedance_terminals_groups_negligible_impedance_equipment(nie_constructor):
+    nb_network = negligible_impedance_equipment_basic_network(lambda mrid: nie_constructor(mrid=mrid))
+
+    nie1 = nb_network.get("nie1")
+    nie2 = nb_network.get("nie2")
+    a0 = nb_network.get("a0")
+    a1 = nb_network.get("a1")
+    a2 = nb_network.get("a2")
+    a3 = nb_network.get("a3")
+    a4 = nb_network.get("a4")
+    a5 = nb_network.get("a5")
+
+    def has_neg_imp(ce) -> bool:
+        if isinstance(ce, Junction) or isinstance(ce, BusbarSection):
+            return True
+        if isinstance(ce, AcLineSegment):
+            return ce.length == 0
+        if isinstance(ce, Switch):
+            return not ce.is_open()
+        else:
+            return False
+
+    await _validate_term_grouping(has_neg_imp, nb_network, "a0_nie1", {nie1, a0, a1}, {*nie1.terminals, *a0.terminals, *a1.terminals}, {get_term(a2, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "a1_nie1", {nie1, a0, a1}, {*nie1.terminals, *a0.terminals, *a1.terminals}, {get_term(a2, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "a1_a2", {nie1, a0, a1}, {*nie1.terminals, *a0.terminals, *a1.terminals}, {get_term(a2, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "a2_nie2", {nie2}, {*nie2.terminals}, {get_term(a2, 2), get_term(a3, 1), get_term(a4, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "a3_nie2", {nie2}, {*nie2.terminals}, {get_term(a2, 2), get_term(a3, 1), get_term(a4, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "a4_nie2", {nie2}, {*nie2.terminals}, {get_term(a2, 2), get_term(a3, 1), get_term(a4, 1)})
+    await _validate_term_grouping(has_neg_imp, nb_network, "a4_a5", set(), set(), {get_term(a4, 2), *a5.terminals})
+
+
+@pytest.mark.asyncio
+async def test_group_negligible_impedance_terminals_end_of_branch_multiple_ec_pec():
+    nb_network = end_of_branch_multiple_ec_pec()
+
+    a2 = nb_network.get("a2")
+    ec = nb_network.get("ec")
+    pec1 = nb_network.get("pec1")
+    pec2 = nb_network.get("pec2")
+
+    def has_neg_imp(_) -> bool:
+        return False
+
+    await _validate_term_grouping(has_neg_imp, nb_network, "a2_ec_pec1_pec2", set(), set(), get_terms({a2: 2, ec: 1, pec1: 1, pec2: 1}))
+
+
+def _get_expected(nb_network, line, pt, es, ec, pec):
     # -- Bus
-    expected_bus_0 = (
-        create_terminal_based_id({next(es.terminals), list(pt.terminals)[0]}),
-        (
-            20000,
-            frozenset(),
-            frozenset({list(es.terminals)[0], list(pt.terminals)[0]}),
-            frozenset(),
-            nb_network
-        )
-    )
-    expected_bus_1 = (
-        create_terminal_based_id({list(pt.terminals)[1], list(line.terminals)[0]}),
-        (
-            400,
-            frozenset(),
-            frozenset({list(line.terminals)[0], list(pt.terminals)[1]}),
-            frozenset(),
-            nb_network
-        )
-    )
-    expected_bus_2 = (
-        create_terminal_based_id({list(line.terminals)[1], next(ec.terminals), next(pec.terminals)}),
-        (
-            400,
-            frozenset(),
-            frozenset({list(ec.terminals)[0], list(line.terminals)[1], list(pec.terminals)[0]}),
-            frozenset(),
-            nb_network
-        )
-    )
+    exp_bb0 = (create_terminal_based_id({next(es.terminals), get_term(pt, 1)}),
+               (20000, frozenset(), frozenset({get_term(es, 1), get_term(pt, 1)}), frozenset(), nb_network))
+    exp_bb1 = (create_terminal_based_id({get_term(pt, 2), get_term(line, 1)}),
+               (400, frozenset(), frozenset({get_term(line, 1), get_term(pt, 2)}), frozenset(), nb_network))
+    exp_bb2 = (create_terminal_based_id({get_term(line, 2), next(ec.terminals), next(pec.terminals)}),
+               (400, frozenset(), frozenset({get_term(ec, 1), get_term(line, 2), get_term(pec, 1)}), frozenset(), nb_network))
 
     # -- Branch
-    expected_branch = (
-        f"tb_{line.mrid}",
-        (
-            (expected_bus_1[1], expected_bus_2[1]),
-            100,
-            frozenset({line}),
-            frozenset({*line.terminals}),
-            frozenset(),
-            nb_network
-        )
-    )
+    exp_branch = (f"tb_{line.mrid}", ((exp_bb1[1], exp_bb2[1]), 100, frozenset({line}), frozenset({*line.terminals}), frozenset(), nb_network))
 
     # -- Transformer
-    expected_end_to_bus_pairs = ((list(pt.ends)[0], expected_bus_0[1]), (list(pt.ends)[1], expected_bus_1[1]))
-    expected_transformer = (
-        f"pt_{pt.mrid}",
-        (
-            pt,
-            expected_end_to_bus_pairs,
-            nb_network
-        )
-    )
+    exp_end_to_bus_pairs = ((list(pt.ends)[0], exp_bb0[1]), (list(pt.ends)[1], exp_bb1[1]))
+    exp_tx = (f"pt_{pt.mrid}", (pt, exp_end_to_bus_pairs, nb_network))
 
     # -- Source
-    expected_energy_source = (
-        f"es_{es.mrid}",
-        (
-            es,
-            expected_bus_0[1],
-            nb_network
-        )
-    )
+    exp_es = (f"es_{es.mrid}", (es, exp_bb0[1], nb_network))
 
     # -- Consumer
-    expected_energy_consumer = (
-        f"ec_{ec.mrid}",
-        (
-            ec,
-            expected_bus_2[1],
-            nb_network
-        )
-    )
+    exp_ec = (f"ec_{ec.mrid}", (ec, exp_bb2[1], nb_network))
 
     # -- PowerElectronicsConnection
-    expected_power_electronics_connection = (
-        f"pec_{pec.mrid}",
-        (
-            pec,
-            expected_bus_2[1],
-            nb_network
-        )
-    )
+    exp_pec = (f"pec_{pec.mrid}", (pec, exp_bb2[1], nb_network))
 
-    return expected_bus_0, expected_bus_1, expected_bus_2, expected_branch, expected_transformer, expected_energy_source, expected_energy_consumer, expected_power_electronics_connection
+    return exp_bb0, exp_bb1, exp_bb2, exp_branch, exp_tx, exp_es, exp_ec, exp_pec
 
 
 def _get_validated_results(result):
-    was_successful = result.was_successful
-    validator = result.validator
-    mappings = result.mappings
-    bb_network = result.network
+    assert result.was_successful is True
+    assert result.validator is not None
+    assert result.mappings is not None
+    assert result.network is not None
 
-    assert was_successful is True
-    assert validator is not None
-    assert mappings is not None
-    assert bb_network is not None
-
-    return was_successful, validator, mappings, bb_network
+    return result.validator, result.mappings, result.network
 
 
 def _get_validated_conducting_equipment(nb_network):
-    assert nb_network is not None
-
     plsi = nb_network.get("plsi")
-    assert plsi is not None
-
     wire_info = nb_network.get("wire_info")
-    assert wire_info is not None
-
     pt_info = nb_network.get("pt_info")
-    assert pt_info is not None
-
     es = nb_network.get("grid_connection")
-    assert es is not None
-
     pt = nb_network.get("transformer")
-    assert pt is not None
-
     line = nb_network.get("line")
-    assert line is not None
-
     ec = nb_network.get("load")
-    assert ec is not None
-
     pec = nb_network.get("pec")
-    assert pec is not None
 
     return plsi, wire_info, pt_info, es, pt, line, ec, pec
 
@@ -200,22 +294,22 @@ def _validate_network(
     expected_energy_consumer,
     expected_power_electronics_connection,
 ):
-    _assert_are_equal(bb_network.bus, {expected_bus_0, expected_bus_1, expected_bus_2})
+    assert bb_network.bus == {expected_bus_0, expected_bus_1, expected_bus_2}
 
     # Branch Comparison
     branch = list(bb_network.branch)[0]
     assert branch[0] == expected_branch[0]
-    _assert_are_equal(set(branch[1][0]), set(expected_branch[1][0]))
+    assert set(branch[1][0]) == set(expected_branch[1][0])
     assert branch[1][1] == expected_branch[1][1]
     assert branch[1][2] == expected_branch[1][2]
     assert branch[1][3] == expected_branch[1][3]
     assert branch[1][4] == expected_branch[1][4]
     assert branch[1][5] == expected_branch[1][5]
 
-    _assert_are_equal(bb_network.transformer, {expected_transformer})
-    _assert_are_equal(bb_network.energy_source, {expected_energy_source})
-    _assert_are_equal(bb_network.energy_consumer, {expected_energy_consumer})
-    _assert_are_equal(bb_network.power_electronics_connection, {expected_power_electronics_connection})
+    assert bb_network.transformer == {expected_transformer}
+    assert bb_network.energy_source == {expected_energy_source}
+    assert bb_network.energy_consumer == {expected_energy_consumer}
+    assert bb_network.power_electronics_connection == {expected_power_electronics_connection}
 
 
 def _validate_mappings(
@@ -279,17 +373,6 @@ def _validate_mappings(
     assert list(mappings.to_bbn.objects[pec.mrid])[0] == expected_power_electronics_connection[1]
 
 
-def _validate_branches_are_equal(a, b):
-    assert a[0][0][0] == b[0][0][0]
-
-    assert a[1] == b[1]
-    assert a[2] == b[2]
-    _assert_are_equal(a[3], b[3])
-    _assert_are_equal(a[4], b[4])
-    _assert_are_equal(a[5], b[5])
-    assert a[6] is b[6]
-
-
 def _validate_bus_mapping(mappings, expected_bus):
     mapped_tn = mappings.to_nbn.topological_nodes[expected_bus[0]]
     _assert_equal_bus(mapped_tn, expected_bus)
@@ -316,467 +399,42 @@ def _assert_equal_branch(terminal_grouping, expected_branch):
     assert terminal_grouping.inner_terminals == expected_branch[1][4]
 
 
-@pytest.mark.asyncio
-async def test_topological_nodes_inside_grouped_lines_do_not_get_created(three_common_lines_network):
-    nb_network = three_common_lines_network
-    creator = TestBusBranchCreator()
-    result = await creator.create(nb_network)
-
-    was_successful, validator, mappings, bb_network = _get_validated_results(result)
-
-    assert len(bb_network.bus) == 2
-    assert len(bb_network.branch) == 1
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize(
-    'single_branch_common_lines_network',
-    [False, True],
-    indirect=True
-)
-async def test_group_common_ac_line_segment_terminals_single_branch(single_branch_common_lines_network):
-    assert single_branch_common_lines_network is not None
-
-    acls1 = single_branch_common_lines_network.get("acls1")
-    assert acls1 is not None
-    acls2 = single_branch_common_lines_network.get("acls2")
-    assert acls2 is not None
-    acls3 = single_branch_common_lines_network.get("acls3")
-    assert acls3 is not None
-    acls4 = single_branch_common_lines_network.get("acls4")
-    assert acls4 is not None
-    acls5 = single_branch_common_lines_network.get("acls5")
-    assert acls5 is not None
-
-    # Validation
-    # acls1, acls2, acls3
-    for a in acls1, acls2, acls3:
-        lines_grouping = await _group_common_ac_line_segment_terminals(a)
-        _assert_are_equal(lines_grouping.conducting_equipment_group, {acls1, acls2, acls3})
-        _assert_are_equal(lines_grouping.inner_terminals, {*acls1.terminals, *acls2.terminals, list(acls3.terminals)[0]})
-        _assert_are_equal(lines_grouping.border_terminals, {list(acls3.terminals)[1]})
-
-    # acls4
-    lines_grouping = await _group_common_ac_line_segment_terminals(acls4)
-    _assert_are_equal(lines_grouping.conducting_equipment_group, {acls4})
-    _assert_are_equal(lines_grouping.inner_terminals, set())
-    _assert_are_equal(lines_grouping.border_terminals, {*acls4.terminals})
-
-    # acls5
-    lines_grouping = await _group_common_ac_line_segment_terminals(acls5)
-    _assert_are_equal(lines_grouping.conducting_equipment_group, {acls5})
-    _assert_are_equal(lines_grouping.inner_terminals, set())
-    _assert_are_equal(lines_grouping.border_terminals, {*acls5.terminals})
-
-
-@pytest.mark.asyncio
-async def test_group_common_ac_line_segment_terminals_multi_branch(multi_branch_common_lines_network):
-    assert multi_branch_common_lines_network is not None
-
-    a0 = multi_branch_common_lines_network.get("a0")
-    assert a0 is not None
-    a1 = multi_branch_common_lines_network.get("a1")
-    assert a1 is not None
-    a2 = multi_branch_common_lines_network.get("a2")
-    assert a2 is not None
-    a3 = multi_branch_common_lines_network.get("a3")
-    assert a3 is not None
-    a4 = multi_branch_common_lines_network.get("a4")
-    assert a4 is not None
-    a5 = multi_branch_common_lines_network.get("a5")
-    assert a5 is not None
-    a6 = multi_branch_common_lines_network.get("a6")
-    assert a6 is not None
-    a7 = multi_branch_common_lines_network.get("a7")
-    assert a7 is not None
-    a8 = multi_branch_common_lines_network.get("a8")
-    assert a8 is not None
-
-    # Validation
-    # a0, a1, a2
-    for a in a0, a1, a2:
-        lines_grouping = await _group_common_ac_line_segment_terminals(a)
-        _assert_are_equal(lines_grouping.conducting_equipment_group, {a0, a1, a2})
-        _assert_are_equal(lines_grouping.inner_terminals, {list(a0.terminals)[0], *a1.terminals, list(a2.terminals)[0]})
-        _assert_are_equal(lines_grouping.border_terminals, {list(a2.terminals)[1]})
-
-    # a3
-    lines_grouping = await _group_common_ac_line_segment_terminals(a3)
-    _assert_are_equal(lines_grouping.conducting_equipment_group, {a3})
-    _assert_are_equal(lines_grouping.inner_terminals, set())
-    _assert_are_equal(lines_grouping.border_terminals, {*a3.terminals})
-
-    # a4, a5
-    for a in a4, a5:
-        lines_grouping = await _group_common_ac_line_segment_terminals(a)
-        _assert_are_equal(lines_grouping.conducting_equipment_group, {a4, a5})
-        _assert_are_equal(lines_grouping.inner_terminals, {list(a4.terminals)[1], list(a5.terminals)[0]})
-        _assert_are_equal(lines_grouping.border_terminals, {list(a4.terminals)[0], list(a5.terminals)[1]})
-
-    # a6, a7
-    for a in a6, a7:
-        lines_grouping = await _group_common_ac_line_segment_terminals(a)
-        _assert_are_equal(lines_grouping.conducting_equipment_group, {a6, a7})
-        _assert_are_equal(lines_grouping.inner_terminals, {list(a6.terminals)[1], *a7.terminals})
-        _assert_are_equal(lines_grouping.border_terminals, {list(a6.terminals)[0]})
-
-    # a8
-    lines_grouping = await _group_common_ac_line_segment_terminals(a8)
-    _assert_are_equal(lines_grouping.conducting_equipment_group, {a8})
-    _assert_are_equal(lines_grouping.inner_terminals, set())
-    _assert_are_equal(lines_grouping.border_terminals, {*a8.terminals})
-
-
-@pytest.mark.asyncio
-async def test_group_common_ac_line_segment_terminals_end_of_branch_multiple_ec_pec(end_of_branch_multiple_ec_pec):
-    assert end_of_branch_multiple_ec_pec is not None
-
-    a1 = end_of_branch_multiple_ec_pec.get("a1")
-    assert a1 is not None
-    a2 = end_of_branch_multiple_ec_pec.get("a2")
-    assert a2 is not None
-    ec = end_of_branch_multiple_ec_pec.get("ec")
-    assert ec is not None
-    pec1 = end_of_branch_multiple_ec_pec.get("pec1")
-    assert pec1 is not None
-    pec2 = end_of_branch_multiple_ec_pec.get("pec2")
-    assert pec2 is not None
-
-    # Validation
-    # a1, a2
-    for a in a1, a2:
-        lines_grouping = await _group_common_ac_line_segment_terminals(a)
-        _assert_are_equal(lines_grouping.conducting_equipment_group, {a1, a2})
-        _assert_are_equal(lines_grouping.inner_terminals, {list(a1.terminals)[1], list(a2.terminals)[0]})
-        _assert_are_equal(lines_grouping.border_terminals, {list(a1.terminals)[0], list(a2.terminals)[1]})
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize(
-    'single_branch_common_lines_network',
-    [False],
-    indirect=True
-)
-async def test_group_negligible_impedance_terminals_single_branch_closed_switch(single_branch_common_lines_network):
-    assert single_branch_common_lines_network is not None
-
-    acls1 = single_branch_common_lines_network.get("acls1")
-    assert acls1 is not None
-    acls2 = single_branch_common_lines_network.get("acls2")
-    assert acls2 is not None
-    acls3 = single_branch_common_lines_network.get("acls3")
-    assert acls3 is not None
-    acls4 = single_branch_common_lines_network.get("acls4")
-    assert acls4 is not None
-    acls5 = single_branch_common_lines_network.get("acls5")
-    assert acls5 is not None
-    sw = single_branch_common_lines_network.get("sw")
-    assert sw is not None
-
-    def has_negligible_impedance(ce: ConductingEquipment) -> bool:
-        if isinstance(ce, Switch):
-            return not ce.is_open()
-        else:
-            return False
-
-    cn = {"_".join(sorted([t.conducting_equipment.mrid for t in cn.terminals])): cn for cn in single_branch_common_lines_network.objects(ConnectivityNode)}
-
-    # Validation
-    # a1_a2
-    terminal_group = await _group_negligible_impedance_terminals(list(cn["acls1_acls2"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_group.conducting_equipment_group, set())
-    _assert_are_equal(terminal_group.inner_terminals, set())
-    _assert_are_equal(terminal_group.border_terminals, {list(acls2.terminals)[0], *acls1.terminals})
-
-    # a2_a3
-    terminal_group = await _group_negligible_impedance_terminals(list(cn["acls2_acls3"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_group.conducting_equipment_group, set())
-    _assert_are_equal(terminal_group.inner_terminals, set())
-    _assert_are_equal(terminal_group.border_terminals, {list(acls2.terminals)[1], list(acls3.terminals)[0]})
-
-    # a3_sw
-    terminal_group = await _group_negligible_impedance_terminals(list(cn["acls3_sw"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_group.conducting_equipment_group, {sw})
-    _assert_are_equal(terminal_group.inner_terminals, {*sw.terminals})
-    _assert_are_equal(terminal_group.border_terminals, {list(acls3.terminals)[1], list(acls4.terminals)[0]})
-
-    # sw_a4
-    terminal_group = await _group_negligible_impedance_terminals(list(cn["acls4_sw"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_group.conducting_equipment_group, {sw})
-    _assert_are_equal(terminal_group.inner_terminals, {*sw.terminals})
-    _assert_are_equal(terminal_group.border_terminals, {list(acls3.terminals)[1], list(acls4.terminals)[0]})
-
-    # a4_a5
-    terminal_group = await _group_negligible_impedance_terminals(list(cn["acls4_acls5"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_group.conducting_equipment_group, set())
-    _assert_are_equal(terminal_group.inner_terminals, set())
-    _assert_are_equal(terminal_group.border_terminals, {list(acls4.terminals)[1], *acls5.terminals})
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize(
-    'single_branch_common_lines_network',
-    [True],
-    indirect=True
-)
-async def test_group_negligible_impedance_terminals_single_branch_open_switch(single_branch_common_lines_network):
-    assert single_branch_common_lines_network is not None
-
-    acls1 = single_branch_common_lines_network.get("acls1")
-    assert acls1 is not None
-    acls2 = single_branch_common_lines_network.get("acls2")
-    assert acls2 is not None
-    acls3 = single_branch_common_lines_network.get("acls3")
-    assert acls3 is not None
-    acls4 = single_branch_common_lines_network.get("acls4")
-    assert acls4 is not None
-    acls5 = single_branch_common_lines_network.get("acls5")
-    assert acls5 is not None
-    sw = single_branch_common_lines_network.get("sw")
-    assert sw is not None
-
-    def has_negligible_impedance(ce: ConductingEquipment) -> bool:
-        if isinstance(ce, Switch):
-            return not ce.is_open()
-        else:
-            return False
-
-    cn = {"_".join(sorted([t.conducting_equipment.mrid for t in cn.terminals])): cn for cn in single_branch_common_lines_network.objects(ConnectivityNode)}
-
-    # Validation
-    # a1_a2
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["acls1_acls2"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, set())
-    _assert_are_equal(terminal_grouping.inner_terminals, set())
-    _assert_are_equal(terminal_grouping.border_terminals, {list(acls2.terminals)[0], *acls1.terminals})
-
-    # a2_a3
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["acls2_acls3"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, set())
-    _assert_are_equal(terminal_grouping.inner_terminals, set())
-    _assert_are_equal(terminal_grouping.border_terminals, {list(acls2.terminals)[1], list(acls3.terminals)[0]})
-
-    # a3_sw
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["acls3_sw"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, set())
-    _assert_are_equal(terminal_grouping.inner_terminals, set())
-    _assert_are_equal(terminal_grouping.border_terminals, {list(acls3.terminals)[1], list(sw.terminals)[0]})
-
-    # sw_a4
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["acls4_sw"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, set())
-    _assert_are_equal(terminal_grouping.inner_terminals, set())
-    _assert_are_equal(terminal_grouping.border_terminals, {list(sw.terminals)[1], list(acls4.terminals)[0]})
-
-    # a4_a5
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["acls4_acls5"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, set())
-    _assert_are_equal(terminal_grouping.inner_terminals, set())
-    _assert_are_equal(terminal_grouping.border_terminals, {list(acls4.terminals)[1], *acls5.terminals})
-
-
-@pytest.mark.asyncio
-async def test_group_negligible_impedance_terminals_multi_branch(multi_branch_common_lines_network):
-    assert multi_branch_common_lines_network is not None
-
-    a0 = multi_branch_common_lines_network.get("a0")
-    assert a0 is not None
-    a1 = multi_branch_common_lines_network.get("a1")
-    assert a1 is not None
-    a2 = multi_branch_common_lines_network.get("a2")
-    assert a2 is not None
-    a3 = multi_branch_common_lines_network.get("a3")
-    assert a3 is not None
-    a4 = multi_branch_common_lines_network.get("a4")
-    assert a4 is not None
-    a5 = multi_branch_common_lines_network.get("a5")
-    assert a5 is not None
-    a6 = multi_branch_common_lines_network.get("a6")
-    assert a6 is not None
-    a7 = multi_branch_common_lines_network.get("a7")
-    assert a7 is not None
-    a8 = multi_branch_common_lines_network.get("a8")
-    assert a8 is not None
-
-    def has_negligible_impedance(ce: ConductingEquipment) -> bool:
-        if isinstance(ce, Switch):
-            return not ce.is_open()
-        else:
-            return False
-
-    cn = {"_".join(sorted([t.conducting_equipment.mrid for t in cn.terminals])): cn for cn in multi_branch_common_lines_network.objects(ConnectivityNode)}
-
-    # Validation
-    # a0_a1
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["a0_a1"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, set())
-    _assert_are_equal(terminal_grouping.inner_terminals, set())
-    _assert_are_equal(terminal_grouping.border_terminals, {list(a1.terminals)[0], *a0.terminals})
-
-    # a1_a2
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["a1_a2"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, set())
-    _assert_are_equal(terminal_grouping.inner_terminals, set())
-    _assert_are_equal(terminal_grouping.border_terminals, {list(a1.terminals)[1], list(a2.terminals)[0]})
-
-    # a2_a3_a6
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["a2_a3_a6"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, set())
-    _assert_are_equal(terminal_grouping.inner_terminals, set())
-    _assert_are_equal(terminal_grouping.border_terminals, {list(a2.terminals)[1], list(a3.terminals)[0], list(a6.terminals)[0]})
-
-    # a3_a4_a8
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["a3_a4_a8"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, set())
-    _assert_are_equal(terminal_grouping.inner_terminals, set())
-    _assert_are_equal(terminal_grouping.border_terminals, {list(a3.terminals)[1], list(a4.terminals)[0], *a8.terminals})
-
-    # a4_a5
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["a4_a5"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, set())
-    _assert_are_equal(terminal_grouping.inner_terminals, set())
-    _assert_are_equal(terminal_grouping.border_terminals, {list(a4.terminals)[1], list(a5.terminals)[0]})
-
-    terminal_grouping = await _group_negligible_impedance_terminals(list(a5.terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, set())
-    _assert_are_equal(terminal_grouping.inner_terminals, set())
-    _assert_are_equal(terminal_grouping.border_terminals, {list(a4.terminals)[1], list(a5.terminals)[0]})
-
-    # a6_a7
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["a6_a7"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, set())
-    _assert_are_equal(terminal_grouping.inner_terminals, set())
-    _assert_are_equal(terminal_grouping.border_terminals, {list(a6.terminals)[1], *a7.terminals})
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize(
-    'negligible_impedance_equipment_basic_network',
-    [
-        lambda mrid: Junction(mrid=mrid),
-        lambda mrid: Disconnector(mrid=mrid),
-        lambda mrid: BusbarSection(mrid=mrid)
-    ],
-    indirect=True
-)
-async def test_group_negligible_impedance_terminals_groups_negligible_impedance_equipment(negligible_impedance_equipment_basic_network):
-    assert negligible_impedance_equipment_basic_network is not None
-
-    nie1 = negligible_impedance_equipment_basic_network.get("nie1")
-    assert nie1 is not None
-    nie2 = negligible_impedance_equipment_basic_network.get("nie2")
-    assert nie2 is not None
-    a0 = negligible_impedance_equipment_basic_network.get("a0")
-    assert a0 is not None
-    a1 = negligible_impedance_equipment_basic_network.get("a1")
-    assert a1 is not None
-    a2 = negligible_impedance_equipment_basic_network.get("a2")
-    assert a2 is not None
-    a3 = negligible_impedance_equipment_basic_network.get("a3")
-    assert a3 is not None
-    a4 = negligible_impedance_equipment_basic_network.get("a4")
-    assert a4 is not None
-    a5 = negligible_impedance_equipment_basic_network.get("a5")
-
-    def has_negligible_impedance(ce: ConductingEquipment) -> bool:
-        if isinstance(ce, Junction) or isinstance(ce, BusbarSection):
-            return True
-        if isinstance(ce, AcLineSegment):
-            return ce.length == 0
-        if isinstance(ce, Switch):
-            return not ce.is_open()
-        else:
-            return False
-
-    cn = {"_".join(sorted([t.conducting_equipment.mrid for t in cn.terminals])): cn for cn in
-          negligible_impedance_equipment_basic_network.objects(ConnectivityNode)}
-
-    # Validation
-    # a0_nie1
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["a0_nie1"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, {nie1, a0, a1})
-    _assert_are_equal(terminal_grouping.inner_terminals, {*nie1.terminals, *a0.terminals, *a1.terminals})
-    _assert_are_equal(terminal_grouping.border_terminals, {list(a2.terminals)[0]})
-
-    # nie1_a1
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["a1_nie1"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, {nie1, a0, a1})
-    _assert_are_equal(terminal_grouping.inner_terminals, {*nie1.terminals, *a0.terminals, *a1.terminals})
-    _assert_are_equal(terminal_grouping.border_terminals, {list(a2.terminals)[0]})
-
-    # a1_a2
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["a1_a2"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, {nie1, a0, a1})
-    _assert_are_equal(terminal_grouping.inner_terminals, {*nie1.terminals, *a0.terminals, *a1.terminals})
-    _assert_are_equal(terminal_grouping.border_terminals, {list(a2.terminals)[0]})
-
-    # a2_nie2
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["a2_nie2"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, {nie2})
-    _assert_are_equal(terminal_grouping.inner_terminals, {*nie2.terminals})
-    _assert_are_equal(terminal_grouping.border_terminals, {list(a2.terminals)[1], list(a3.terminals)[0], list(a4.terminals)[0]})
-
-    # a3_nie2
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["a3_nie2"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, {nie2})
-    _assert_are_equal(terminal_grouping.inner_terminals, {*nie2.terminals})
-    _assert_are_equal(terminal_grouping.border_terminals, {list(a2.terminals)[1], list(a3.terminals)[0], list(a4.terminals)[0]})
-
-    # a4_nie2
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["a4_nie2"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, {nie2})
-    _assert_are_equal(terminal_grouping.inner_terminals, {*nie2.terminals})
-    _assert_are_equal(terminal_grouping.border_terminals, {list(a2.terminals)[1], list(a3.terminals)[0], list(a4.terminals)[0]})
-
-    # a4_a5
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["a4_a5"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, set())
-    _assert_are_equal(terminal_grouping.inner_terminals, set())
-    _assert_are_equal(terminal_grouping.border_terminals, {list(a4.terminals)[1], *a5.terminals})
-
-
-@pytest.mark.asyncio
-async def test_group_negligible_impedance_terminals_end_of_branch_multiple_ec_pec(end_of_branch_multiple_ec_pec):
-    assert end_of_branch_multiple_ec_pec is not None
-
-    a1 = end_of_branch_multiple_ec_pec.get("a1")
-    assert a1 is not None
-    a2 = end_of_branch_multiple_ec_pec.get("a2")
-    assert a2 is not None
-    ec = end_of_branch_multiple_ec_pec.get("ec")
-    assert ec is not None
-    pec1 = end_of_branch_multiple_ec_pec.get("pec1")
-    assert pec1 is not None
-    pec2 = end_of_branch_multiple_ec_pec.get("pec2")
-    assert pec2 is not None
-
-    def has_negligible_impedance(ce: ConductingEquipment) -> bool:
-        return False
-
-    cn = {"_".join(sorted([t.conducting_equipment.mrid for t in cn.terminals])): cn for cn in end_of_branch_multiple_ec_pec.objects(ConnectivityNode)}
-
-    # Validation
-    # a2_ec
-    terminal_grouping = await _group_negligible_impedance_terminals(list(cn["a2_ec_pec1_pec2"].terminals)[0], has_negligible_impedance)
-    _assert_are_equal(terminal_grouping.conducting_equipment_group, set())
-    _assert_are_equal(terminal_grouping.inner_terminals, set())
-    _assert_are_equal(terminal_grouping.border_terminals, {list(a2.terminals)[1], list(ec.terminals)[0], list(pec1.terminals)[0], list(pec2.terminals)[0]})
-
-
-def _assert_are_equal(a: Union[Set, FrozenSet], b: Union[Set, FrozenSet]):
-    diff = a ^ b
-    assert not diff, f"Sets {a} and {b} are different"
-
-
-def _assert_sets_of_sets_are_equal(a: Union[Set, FrozenSet], b: Union[Set, FrozenSet]):
-    b_c = b.copy()
-    for cl in a:
-        matched_group = None
-        for ex_g in b_c:
-            diff = cl ^ ex_g
-            if not diff:
-                matched_group = ex_g
-                break
-        if matched_group is not None:
-            b_c.remove(matched_group)
-    assert len(b_c) == 0, f"Sets are not equal. Number of non-matching elements: {len(b_c)}"
+async def _validate_line_grouping(
+    expected_acls: Set[AcLineSegment],
+    expected_inner: Set[Terminal],
+    expected_border: Set[Terminal]
+):
+    for a in expected_acls:
+        grouping = await _group_common_ac_line_segment_terminals(a)
+        assert grouping.conducting_equipment_group == expected_acls
+        assert grouping.inner_terminals == expected_inner
+        assert grouping.border_terminals == expected_border
+
+
+async def _validate_term_grouping(
+    has_negligible_impedance: Callable[[ConductingEquipment], bool],
+    nb_network: NetworkService,
+    mrid: str,
+    expected_ce: Set[ConductingEquipment],
+    expected_inner: Set[Terminal],
+    expected_border: Set[Terminal]
+):
+    ce = nb_network.get(mrid, ConductingEquipment, None)
+    if ce is not None:
+        terminal = get_term(ce, 1)
+    else:
+        cn = {"_".join(sorted([t.conducting_equipment.mrid for t in cn.terminals])): cn for cn in nb_network.objects(ConnectivityNode)}
+        terminal = list(cn[mrid].terminals)[0]
+
+    grouping = await _group_negligible_impedance_terminals(terminal, has_negligible_impedance)
+    assert grouping.conducting_equipment_group == expected_ce
+    assert grouping.inner_terminals == expected_inner
+    assert grouping.border_terminals == expected_border
+
+
+def get_term(ce: ConductingEquipment, term: int) -> Terminal:
+    return ce.get_terminal_by_sn(term)
+
+
+def get_terms(request: Dict[ConductingEquipment, int]) -> Set[Terminal]:
+    return {get_term(ce, term) for ce, term in request.items()}
